have grunt copy patternfly images along with fonts and less files
When trying to consume rcue via project dependency, and using
webpack to pull assets to an application from a node_module, all
of the images need to be present or the webpack will fail.  Copy
all the images across on each update to make sure changes and
additions get included automatically.